implement 'projects document' impled project endpoint
