+++ docs/handlers/projects/_lib/common-constants.mjs.html
@@ -1,0 +1,39 @@
+<html>
+  <head>
+    <title>common-constants.mjs</title>
+
+    <style>
+      body {
+        text-align: left;
+      }
+
+      pre {
+        text-align: left;
+        display: inline-block;
+        max-width: 120em;
+        margin: 1em 1em;
+        padding: 1em 1em;
+        border-top: solid blue 1px;
+        border-right: solid blue 1px;
+        border-bottom: solid blue 1px;
+        white-space: pre-wrap;
+      }
+    </style>
+
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.7.0/styles/base16/google-light.min.css">
+  </head>
+  <body>
+    <pre><code>
+<span class="hljs-keyword">const</span> <span class="hljs-variable constant_">DEFAULT_LICENSE</span> = <span class="hljs-string">&#x27;UNLICENSED&#x27;</span>
+<span class="hljs-keyword">const</span> <span class="hljs-variable constant_">DEFAULT_VERSION</span> = <span class="hljs-string">&#x27;1.0.0-alpha.0&#x27;</span>
+<span class="hljs-keyword">const</span> <span class="hljs-variable constant_">GITHUB_REPO_KEY</span> = <span class="hljs-string">&#x27;core.github.ORG_NAME&#x27;</span>
+
+<span class="hljs-keyword">export</span> {
+  <span class="hljs-variable constant_">DEFAULT_LICENSE</span>,
+  <span class="hljs-variable constant_">DEFAULT_VERSION</span>,
+  <span class="hljs-variable constant_">GITHUB_REPO_KEY</span>
+}
+
+    </code></pre>
+  </body>
+</html>
