+++ docs/handlers/projects/_lib/common-project-path-parameters.mjs.html
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <title>index.js</title>
+    <title>common-project-path-parameters.mjs</title>
 
     <style>
       body {
@@ -24,13 +24,14 @@
   </head>
   <body>
     <pre><code>
-<span class="hljs-comment">// import { handlers as orgHandlers } from &#x27;./orgs&#x27;</span>
-<span class="hljs-comment">// import { handlers as projectHandlers } from &#x27;./projects&#x27;</span>
-<span class="hljs-keyword">import</span> { handlers } <span class="hljs-keyword">from</span> <span class="hljs-string">&#x27;./projects&#x27;</span>
+<span class="hljs-keyword">const</span> commonProjectPathParameters = [
+  {
+    name        : <span class="hljs-string">&#x27;projectPath&#x27;</span>,
+    description : <span class="hljs-string">&#x27;Path to local project to work with. Can be used to specify projects in non-default locations. Otherwise, the process will look in the liq playground under the `&amp;lt;orgKey&amp;gt;/&amp;lt;localProjectName&amp;gt;`.&#x27;</span>
+  }
+]
 
-<span class="hljs-comment">// const handlers = [...orgHandlers].push(...projectHandlers)</span>
-
-<span class="hljs-keyword">export</span> { handlers }
+<span class="hljs-keyword">export</span> { commonProjectPathParameters }
 
     </code></pre>
   </body>
