+++ docs/handlers/projects/document-implied.mjs.html
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <title>document.js</title>
+    <title>document-implied.mjs</title>
 
     <style>
       body {
@@ -24,19 +24,24 @@
   </head>
   <body>
     <pre><code>
+<span class="hljs-keyword">import</span> createError <span class="hljs-keyword">from</span> <span class="hljs-string">&#x27;http-errors&#x27;</span>
+
+<span class="hljs-keyword">import</span> { determineImpliedProject } <span class="hljs-keyword">from</span> <span class="hljs-string">&#x27;@liquid-labs/liq-projects-lib&#x27;</span>
+
 <span class="hljs-keyword">import</span> { doDocument, getDocumentEndpointParameters } <span class="hljs-keyword">from</span> <span class="hljs-string">&#x27;./_lib/document-lib&#x27;</span>
 
-<span class="hljs-keyword">const</span> paths = [
-  [<span class="hljs-string">&#x27;projects&#x27;</span>, <span class="hljs-string">&#x27;:orgKey&#x27;</span>, <span class="hljs-string">&#x27;:localProjectName&#x27;</span>, <span class="hljs-string">&#x27;document&#x27;</span>],
-  [<span class="hljs-string">&#x27;orgs&#x27;</span>, <span class="hljs-string">&#x27;:orgKey&#x27;</span>, <span class="hljs-string">&#x27;projects&#x27;</span>, <span class="hljs-string">&#x27;:localProjectName&#x27;</span>, <span class="hljs-string">&#x27;document&#x27;</span>]
-]
+<span class="hljs-keyword">const</span> path = [ <span class="hljs-string">&#x27;projects&#x27;</span>, <span class="hljs-string">&#x27;document&#x27;</span> ]
 
 <span class="hljs-keyword">const</span> { help, method, parameters } = <span class="hljs-title function_">getDocumentEndpointParameters</span>({ workDesc : <span class="hljs-string">&#x27;named&#x27;</span> })
 
 <span class="hljs-keyword">const</span> <span class="hljs-title function_">func</span> = (<span class="hljs-params">{ model, reporter }</span>) =&gt; <span class="hljs-title function_">async</span>(req, res) =&gt; {
   reporter = reporter.<span class="hljs-title function_">isolate</span>()
 
-  <span class="hljs-keyword">const</span> { localProjectName, orgKey } = req.<span class="hljs-property">vars</span>
+  <span class="hljs-keyword">const</span> cwd = req.<span class="hljs-title function_">get</span>(<span class="hljs-string">&#x27;X-CWD&#x27;</span>)
+  <span class="hljs-keyword">if</span> (cwd === <span class="hljs-literal">undefined</span>) {
+    <span class="hljs-keyword">throw</span> createError.<span class="hljs-title class_">BadRequest</span>(<span class="hljs-string">&quot;Called &#x27;work document&#x27; with implied work, but &#x27;X-CWD&#x27; header not found.&quot;</span>)
+  }
+  <span class="hljs-keyword">const</span> [orgKey, localProjectName] = <span class="hljs-title function_">determineImpliedProject</span>({ currDir : cwd }).<span class="hljs-title function_">split</span>(<span class="hljs-string">&#x27;/&#x27;</span>)
 
   <span class="hljs-keyword">await</span> <span class="hljs-title function_">doDocument</span>({ orgKey, localProjectName, req, res })
 }
@@ -46,7 +51,7 @@
   help,
   method,
   parameters,
-  paths
+  path
 }
 
     </code></pre>
