+++ docs/handlers/projects/releases/index.js.html
@@ -24,11 +24,9 @@
   </head>
   <body>
     <pre><code>
-<span class="hljs-comment">// import { handlers as orgHandlers } from &#x27;./orgs&#x27;</span>
-<span class="hljs-comment">// import { handlers as projectHandlers } from &#x27;./projects&#x27;</span>
-<span class="hljs-keyword">import</span> { handlers } <span class="hljs-keyword">from</span> <span class="hljs-string">&#x27;./projects&#x27;</span>
+<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> prepareAndPublishHandler <span class="hljs-keyword">from</span> <span class="hljs-string">&#x27;./prepare-and-publish&#x27;</span>
 
-<span class="hljs-comment">// const handlers = [...orgHandlers].push(...projectHandlers)</span>
+<span class="hljs-keyword">const</span> handlers = [prepareAndPublishHandler]
 
 <span class="hljs-keyword">export</span> { handlers }
