+++ docs/index.html
@@ -5,25 +5,85 @@
     <body>
       <ul>
         <li>
-          <a href=/handlers/index.js.html>/handlers/index.js</a>
+          <a href=handlers/index.js.html>handlers/index.js</a>
         </li>
         <li>
-          <a href=/handlers/projects/_lib/get-files.mjs.html>/handlers/projects/_lib/get-files.mjs</a>
+          <a href=handlers/projects/_lib/archive-lib.mjs.html>handlers/projects/_lib/archive-lib.mjs</a>
         </li>
         <li>
-          <a href=/handlers/projects/_lib/get-package-data.mjs.html>/handlers/projects/_lib/get-package-data.mjs</a>
+          <a href=handlers/projects/_lib/common-constants.mjs.html>handlers/projects/_lib/common-constants.mjs</a>
         </li>
         <li>
-          <a href=/handlers/projects/_lib/test/get-files.test.js.html>/handlers/projects/_lib/test/get-files.test.js</a>
+          <a href=handlers/projects/_lib/common-project-path-parameters.mjs.html>handlers/projects/_lib/common-project-path-parameters.mjs</a>
         </li>
         <li>
-          <a href=/handlers/projects/index.js.html>/handlers/projects/index.js</a>
+          <a href=handlers/projects/_lib/detail-lib.mjs.html>handlers/projects/_lib/detail-lib.mjs</a>
         </li>
         <li>
-          <a href=/handlers/projects/document.js.html>/handlers/projects/document.js</a>
+          <a href=handlers/projects/_lib/document-lib.mjs.html>handlers/projects/_lib/document-lib.mjs</a>
         </li>
         <li>
-          <a href=/index.js.html>/index.js</a>
+          <a href=handlers/projects/_lib/destroy-lib.mjs.html>handlers/projects/_lib/destroy-lib.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/_lib/common-project-setup-parameters.mjs.html>handlers/projects/_lib/common-project-setup-parameters.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/_lib/get-files.mjs.html>handlers/projects/_lib/get-files.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/_lib/test/get-files.test.js.html>handlers/projects/_lib/test/get-files.test.js</a>
+        </li>
+        <li>
+          <a href=handlers/projects/_lib/get-package-data.mjs.html>handlers/projects/_lib/get-package-data.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/archive-implied.mjs.html>handlers/projects/archive-implied.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/archive.mjs.html>handlers/projects/archive.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/create.mjs.html>handlers/projects/create.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/detail-implied.mjs.html>handlers/projects/detail-implied.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/detail.mjs.html>handlers/projects/detail.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/destroy.mjs.html>handlers/projects/destroy.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/destroy-implied.mjs.html>handlers/projects/destroy-implied.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/document-implied.mjs.html>handlers/projects/document-implied.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/index.js.html>handlers/projects/index.js</a>
+        </li>
+        <li>
+          <a href=handlers/projects/document.js.html>handlers/projects/document.js</a>
+        </li>
+        <li>
+          <a href=handlers/projects/releases/prepare-and-publish.mjs.html>handlers/projects/releases/prepare-and-publish.mjs</a>
+        </li>
+        <li>
+          <a href=handlers/projects/releases/index.js.html>handlers/projects/releases/index.js</a>
+        </li>
+        <li>
+          <a href=handlers/projects/rename.mjs.html>handlers/projects/rename.mjs</a>
+        </li>
+        <li>
+          <a href=index.js.html>index.js</a>
+        </li>
+        <li>
+          <a href=handlers/projects/setup.mjs.html>handlers/projects/setup.mjs</a>
+        </li>
+        <li>
+          <a href=setup.mjs.html>setup.mjs</a>
         </li>
       </ul>
     </body>
