+++ docs/index.js.html
@@ -4,8 +4,7 @@
 
     <style>
       body {
-        width: 100%;
-        text-align: center;
+        text-align: left;
       }
 
       pre {
@@ -17,6 +16,7 @@
         border-top: solid blue 1px;
         border-right: solid blue 1px;
         border-bottom: solid blue 1px;
+        white-space: pre-wrap;
       }
     </style>
 
@@ -25,6 +25,8 @@
   <body>
     <pre><code>
 <span class="hljs-keyword">export</span> * <span class="hljs-keyword">from</span> <span class="hljs-string">&#x27;./handlers&#x27;</span>
+<span class="hljs-keyword">export</span> * <span class="hljs-keyword">from</span> <span class="hljs-string">&#x27;./setup&#x27;</span>
+
     </code></pre>
   </body>
 </html>
