+++ src/handlers/projects/index.js
@@ -8,6 +8,7 @@ import * as destroyImpliedHandler from './destroy-implied'
 import * as detailHandler from './detail'
 import * as detailImpliedHandler from './detail-implied'
 import * as documentHandler from './document'
+import * as documentImpliedHandler from './document-implied'
 import * as renameHandler from './rename'
 import * as setupHandler from './setup'
 
@@ -20,6 +21,7 @@ handlers.push(
   detailHandler,
   detailImpliedHandler,
   documentHandler,
+  documentImpliedHandler,
   renameHandler,
   setupHandler
 )
